+++ mreg_cli/utilities/api.py
@@ -11,7 +11,8 @@ import os
 import re
 import sys
 from contextvars import ContextVar
-from typing import Any, Literal, NoReturn, TypeVar, get_origin, overload
+from functools import wraps
+from typing import Any, Callable, Literal, NoReturn, ParamSpec, TypeVar, get_origin, overload
 from urllib.parse import urljoin
 from uuid import uuid4
 
@@ -22,6 +23,7 @@ from requests import Response
 
 from mreg_cli.__about__ import __version__
 from mreg_cli.api.errors import parse_mreg_error
+from mreg_cli.cache import get_cache
 from mreg_cli.config import MregCliConfig
 from mreg_cli.exceptions import (
     APINotOk,
@@ -335,6 +337,25 @@ def get(path: str, params: QueryParams | None = ...) -> Response: ...
 
 def get(path: str, params: QueryParams | None = None, ok404: bool = False) -> Response | None:
     """Make a standard get request."""
+    return _do_get(path, params, ok404)
+
+
+def _do_get(path: str, params: QueryParams | None = None, ok404: bool = False) -> Response | None:
+    """Perform a GET request.
+
+    Separated out from get(), so that we can patch this function with a memoized version
+    without affecting other modules that do `from mreg_cli.utilities.api import get`, as
+    they would then operate on the unpatched version instead of the modified one,
+    since their `get` symbol differs from the `get` symbol in this module
+    in a scenario where `get` itself is patched _after_ it is imported elsewhere.
+
+    The caching module can modify this function instead of `get`,
+    allowing other modules to be oblivious of the caching behavior, and freely
+    import `get` into their namespaces.
+
+    Yes... Patching sucks when you have other modules that import scoped symbols
+    into their own namespace.
+    """
     if params is None:
         params = {}
     return _request_wrapper("get", path, params=params, ok404=ok404)
@@ -591,6 +612,26 @@ def get_typed(
         return adapter.validate_json(resp.text)
 
 
+P = ParamSpec("P")
+
+
+def clear_cache(f: Callable[P, T]) -> Callable[P, T]:
+    """Clear the API cache after running the function."""
+
+    @wraps(f)
+    def wrapper(*args: P.args, **kwargs: P.kwargs) -> T:
+        result = f(*args, **kwargs)
+        try:
+            cache = get_cache()
+            cache.cache.evict("api")  # does not reset stats
+        except Exception as e:
+            logger.error(f"Error clearing cache: {e}")
+        return result
+
+    return wrapper
+
+
+@clear_cache
 def post(path: str, params: QueryParams | None = None, **kwargs: Any) -> Response | None:
     """Use requests to make a post request. Assumes that all kwargs are data fields."""
     if params is None:
@@ -598,6 +639,7 @@ def post(path: str, params: QueryParams | None = None, **kwargs: Any) -> Respons
     return _request_wrapper("post", path, params=params, **kwargs)
 
 
+@clear_cache
 def patch(path: str, params: QueryParams | None = None, **kwargs: Any) -> Response | None:
     """Use requests to make a patch request. Assumes that all kwargs are data fields."""
     if params is None:
@@ -605,6 +647,7 @@ def patch(path: str, params: QueryParams | None = None, **kwargs: Any) -> Respon
     return _request_wrapper("patch", path, params=params, **kwargs)
 
 
+@clear_cache
 def delete(path: str, params: QueryParams | None = None) -> Response | None:
     """Use requests to make a delete request."""
     if params is None: